# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AssignPermissionRequest:
    properties:
      permission_ids:
        items:
          type: integer
        type: array
    required:
    - permission_ids
    type: object
...
  dto.FraudLogResponse:
    properties:
      ad_account_id:
        type: string
      ad_id:
        type: string
      adset_id:
        type: string
      brand_id:
        type: integer
      campaign_id:
        type: string
      created_at:
        type: string
      creative_id:
        type: string
      detected_at:
        type: string
      event_type:
        type: string
      id:
        type: integer
      matched_rule_id:
        type: integer
      message:
        type: string
      new_value:
        type: string
      old_value:
        type: string
      resolved_at:
        type: string
      severity:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.PermissionRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  dto.PermissionResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.RoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  dto.RoleResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/dto.PermissionResponse'
        type: array
    type: object
...
  response.Meta:
    properties:
      last_page:
        type: integer
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  response.PaginationResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        example: true
        type: boolean
    type: object
...
  response.SuccessResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8888
info:
  contact: {}
  description: ERP Ready Backend for Meta Ads & Business Management
  title: Ads Backend API
  version: "1.0"
paths:
...
  /fraud-logs:
    get:
      consumes:
      - application/json
      description: Get a list of fraud logs with pagination and filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Brand ID
        in: query
        name: brand_id
        type: integer
      - description: Ad Account ID
        in: query
        name: ad_account_id
        type: string
      - description: Campaign ID
        in: query
        name: campaign_id
        type: string
      - description: Creative ID
        in: query
        name: creative_id
        type: string
      - description: Severity
        in: query
        name: severity
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: Date Start
        in: query
        name: date_start
        type: string
      - description: Date Stop
        in: query
        name: date_stop
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.FraudLogResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get fraud logs
      tags:
      - Fraud Logs
  /fraud-logs/{id}:
    get:
      consumes:
      - application/json
      description: Get a fraud log by its ID
      parameters:
      - description: Fraud Log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.FraudLogResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get fraud log by ID
      tags:
      - Fraud Logs
  /fraud-logs/{id}/resolve:
    put:
      consumes:
      - application/json
      description: Mark a fraud log as resolved
      parameters:
      - description: Fraud Log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.FraudLogResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Resolve a fraud log
      tags:
      - Fraud Logs
...
  /permissions:
    get:
      consumes:
      - application/json
      description: Retrieve all permissions
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PermissionResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get All Permissions
      tags:
      - Permission Management
    post:
      consumes:
      - application/json
      description: Create a new permission
      parameters:
      - description: Create Permission Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Create Permission
      tags:
      - Permission Management
  /permissions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a permission by ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Delete Permission
      tags:
      - Permission Management
    get:
      consumes:
      - application/json
      description: Retrieve a permission by its ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PermissionResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get Permission by ID
      tags:
      - Permission Management
    put:
      consumes:
      - application/json
      description: Update a permission by ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Permission Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Update Permission
      tags:
      - Permission Management
  /roles:
    get:
      consumes:
      - application/json
      description: Retrieve all roles with permissions
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RoleResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get All Roles
      tags:
      - Role Management
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: Create Role Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Create Role
      tags:
      - Role Management
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Delete Role
      tags:
      - Role Management
    get:
      consumes:
      - application/json
      description: Retrieve a role by its ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get Role by ID
      tags:
      - Role Management
    put:
      consumes:
      - application/json
      description: Update a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Role Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Update Role
      tags:
      - Role Management
  /roles/{id}/permissions:
    post:
      consumes:
      - application/json
      description: Assign permissions to a role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Assign Permission Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AssignPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Assign Permissions
      tags:
      - Role Management
...
securityDefinitions:
  BearerAuth:
    description: Tuliskan 'Bearer ' diikuti dengan token JWT Anda.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"